Project/backend/games_platform/pyproject.toml
[tool.poetry]
name = "games_platform"
version = "0.1.0"
description = ""
authors = ["Bekmaganbetov Zhanbolat <180107180@stu.sdu.edu.kz>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1"
Django = "^4.0.3"
django-filter = "^21.1"
flake8 = "^4.0.1"
isort = "^5.10.1"
autoflake = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
